+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added uv support
 ### Fixed
 - KpiPageTemplate.get_kpi_grid fixed bug where column show NaN if scenario name contains a dot.
+- setup.py and pyproject.toml
 
 ## [0.1.2.4b3] - 2026-02-16
 ### Added